Add BC layer for DocumentRepository deprecation
