+++ tests/Documents/BaseCategoryRepository.php
@@ -2,6 +2,8 @@
 
 namespace Documents;
 
-use Doctrine\ODM\MongoDB\DocumentRepository;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 
-class BaseCategoryRepository extends DocumentRepository {}
+class BaseCategoryRepository extends DocumentRepository
+{
+}
