+++ tests/Documents/CustomRepository/Repository.php
@@ -2,6 +2,8 @@
 
 namespace Documents\CustomRepository;
 
-use Doctrine\ODM\MongoDB\DocumentRepository;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 
-class Repository extends DocumentRepository {}
+class Repository extends DocumentRepository
+{
+}
