+++ tools/sandbox/Documents/UserRepository.php
@@ -2,7 +2,7 @@
 
 namespace Documents;
 
-use Doctrine\ODM\MongoDB\DocumentRepository;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 
 /**
  * UserRepository
